make gcc shut up a bit more
git-svn-id: file:///home/lennart/svn/public/pulseaudio/trunk@1987 fefdeb5f-60dc-0310-8127-8f9354f1896f

+++ src/pulsecore/core-util.c
@@ -1522,7 +1522,7 @@ char *pa_readlink(const char *p) {
             return NULL;
         }
 
-        if (n < l-1) {
+        if ((size_t) n < l-1) {
             c[l-1] = 0;
             return c;
         }
